separate UIView constraint category, utilize it in the AboutViewController
