+++ WMF Framework/UIView+Constraints.swift
@@ -1,0 +1,39 @@
+import UIKit
+
+extension UIView {
+    fileprivate func wmf_deactivateHeightConstraint() {
+        let containerViewHeight = constraints.first(where: {constraint in
+            return (constraint.firstAttribute == .height)
+        })
+        containerViewHeight?.isActive = false
+    }
+    
+    /// Configures the view to expand vertically dynamically to encompass the height of the subview even if the subview's height changes as a result of AutoLayout changes. Details: http://stackoverflow.com/a/35431534/135557
+    @objc public func wmf_addHeightDeterminingSubviewWithConstraintsToEdges(_ subview: UIView) {
+        wmf_addHeightDeterminingSubview(subview, withConstraintsToEdgesWithInsets: .zero)
+    }
+    
+    @objc public func wmf_addHeightDeterminingSubview(_ subview: UIView, withConstraintsToEdgesWithInsets insets: UIEdgeInsets, priority: UILayoutPriority = .required) {
+        wmf_deactivateHeightConstraint()
+        subview.translatesAutoresizingMaskIntoConstraints = false
+        wmf_addSubview(subview, withConstraintsToEdgesWithInsets: insets, priority: priority)
+    }
+    
+    @objc public func wmf_addSubviewWithConstraintsToEdges(_ subview: UIView) {
+        wmf_addSubview(subview, withConstraintsToEdgesWithInsets: .zero)
+    }
+    
+    @objc public func wmf_addSubview(_ subview: UIView, withConstraintsToEdgesWithInsets insets: UIEdgeInsets, priority: UILayoutPriority = .required) {
+        subview.frame = UIEdgeInsetsInsetRect(bounds, insets)
+        addSubview(subview)
+        let topConstraint = topAnchor.constraint(equalTo: subview.topAnchor, constant: insets.top)
+        topConstraint.priority = priority
+        let bottomConstraint = bottomAnchor.constraint(equalTo: subview.bottomAnchor, constant: insets.bottom)
+        topConstraint.priority = priority
+        let leftConstraint = leftAnchor.constraint(equalTo: subview.leftAnchor, constant: insets.left)
+        topConstraint.priority = priority
+        let rightConstraint = rightAnchor.constraint(equalTo: subview.rightAnchor, constant: insets.right)
+        topConstraint.priority = priority
+        addConstraints([topConstraint, bottomConstraint, leftConstraint, rightConstraint])
+    }
+}
